Create LICENSE
Signed-off-by: Dan Kohn <dan@dankohn.com>

You specify the ISC License in package.json. But, it is important to also have a LICENSE file so that potential users can easily understand the license and it is included in a legally binding way.

GitHub uses a library called Licensee to identify a project's license type. It shows this information in the status bar and via the API if it can unambiguously identify the license. This commit creates a LICENSE file so that it contains the full text of the ISC license. This allows Licensee to successfully identify the
license type of Shep as ISC.

Another goal is to have your license show up correctly here:

https://landscape.cncf.io/grouping=landscape&landscape=serverless&selected=shep